Try to guess freezes
Freezes can't be exactly connected to their owner because there's no
unique identifier to connect the old CDP which had an owner and the new
one which doesn't.

This more complex algo tries to best guess when freezes happen, trying
to connect CDPs based on the debt, collateral and iAsset properties.
It's by GPT-4 and I don't care to fully understand the logic. If it
doesn't work, I'll drop freeze event support from this bot.