Detect CDP merges
This is a merged CDP in /cdps:

{
  "output_hash": "085fdeab7c5e061cd5118080bcc1b633fe790c87831087fda0cb1aaf64b54f13",
  "output_index": 0,
  "owner": "",
  "asset": "iBTC",
  "collateralAmount": 359225273605,
  "mintedAmount": 3389265
}

Its owner is neither a 56 char hex string like a normal CDP's, nor
`null` like a frozen CDP's. Yet another surprise from the API...

When this empty string transitions to `null`, that's when I chose to
emit the merge event. I have no idea what "" transitioning to `null` is
supposed to actually mean, but otherwise there'd have to be global state
which keeps track of which empty string CDP we already notified for,
instead of the pure function diffing of two /cdps replies that
generate_cdp_events() does.